requirements.txt
# Azure SDK
azure-identity>=1.15.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-policyinsights>=1.0.0
azure-monitor-query>=1.2.0

# Web app
flask>=3.0.0
stripe>=8.0.0

# Testing
pytest>=8.0.0

# Utilities
python-dotenv>=1.0.0
